Implement the dead reckoning test strategy
      * TODO - read topnav config before launching the strategy
      Or find other way to configure the strategy dynamically

      * Implement the "around the circle" strategy

+++ topnav_driving_strategies/src/main/java/com/github/rosjava/topnav_rosjava_kasptom/topnav_driving_strategies/deadReckoning/IManeuver.java
@@ -7,5 +7,5 @@ public interface IManeuver {
 
     WheelsVelocities getNextVelocity(long timeMs);
 
-    boolean isFinished(double timeMs);
+    boolean isFinished(long timeMs);
 }
