+++ topnav_driving_strategies/src/main/java/com/github/rosjava/topnav_rosjava_kasptom/topnav_driving_strategies/deadReckoning/maneuver/AroundCircleManeuver.java
@@ -18,7 +18,7 @@ public class AroundCircleManeuver implements IManeuver {
     }
 
     @Override
-    public boolean isFinished(double timeMs) {
+    public boolean isFinished(long timeMs) {
         return false;
     }
 }
