+++ topnav_driving_strategies/src/main/java/com/github/rosjava/topnav_rosjava_kasptom/topnav_driving_strategies/deadReckoning/maneuver/StraightLineManeuver.java
@@ -35,7 +35,7 @@ public class StraightLineManeuver implements IManeuver {
     }
 
     @Override
-    public boolean isFinished(double timeMs) {
+    public boolean isFinished(long timeMs) {
         return maneuverDurationMs < timeMs;
     }
